fix: Updating upgrade factories for 8.6 to 8.7
