--- optimize/upgrade/src/test/java/io/camunda/optimize/upgrade/migrate313to86/DeleteProcessInstanceArchiveIndex86IT.java.out
@@ -1,39 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-// TODO recreate C8 IT equivalent of this with #13337
-// package io.camunda.optimize.upgrade.migrate313to86;
-//
-// import static org.assertj.core.api.Assertions.assertThat;
-//
-// import java.util.List;
-// import lombok.SneakyThrows;
-// import org.junit.jupiter.api.Test;
-//
-// public class DeleteProcessInstanceArchiveIndex86IT extends AbstractUpgrade86IT {
-//
-//  @Test
-//  @SneakyThrows
-//  public void deleteProcessInstanceArchiveIndex() {
-//    // given
-//    List<String> processInstanceArchiveIndexList =
-//        getPrefixAwareClient().getAllIndexNames().stream()
-//            .filter(indexName -> indexName.contains("process-instance-archive-"))
-//            .toList();
-//    assertThat(processInstanceArchiveIndexList).hasSize(2);
-//
-//    // when
-//    performUpgrade();
-//
-//    // then
-//    List<String> processInstanceArchiveIndexListPostUpgrade =
-//        getPrefixAwareClient().getAllIndexNames().stream()
-//            .filter(indexName -> indexName.contains("process-instance-archive-"))
-//            .toList();
-//    assertThat(processInstanceArchiveIndexListPostUpgrade).isEmpty();
-//  }
-// }
